requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
django-ckeditor==6.3.0
django-environ==0.8.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-taggit==2.1.0
django-widget-tweaks==1.4.12
gunicorn==20.1.0
Pillow==9.1.0
psycopg2==2.9.3
six==1.16.0
sqlparse==0.4.2
whitenoise==6.0.0
